requirements.txt
# -*- coding: utf-8 -*-
#
# This file is part of Zenodo.
# Copyright (C) 2015-2019 CERN.
#
# Zenodo is free software; you can redistribute it
# and/or modify it under the terms of the GNU General Public License as
# published by the Free Software Foundation; either version 2 of the
# License, or (at your option) any later version.
#
# Zenodo is distributed in the hope that it will be
# useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
# General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with Zenodo; if not, write to the
# Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston,
# MA 02111-1307, USA.
#
# In applying this license, CERN does not
# waive the privileges and immunities granted to it by virtue of its status
# as an Intergovernmental Organization or submit itself to any jurisdiction.

# Pinned base requirements
-r requirements.pinned.txt

# Requirements not on PyPI
-e git+https://github.com/zenodo/zenodo-accessrequests.git#egg=zenodo-accessrequests

# PyPI requirements
alembic==1.0.8
amqp==2.4.2
angular-gettext-babel==0.3
aniso8601==6.0.0
arrow==0.13.1
asn1crypto==1.3.0
attrs==19.1.0
autosemver==0.5.3
Babel==2.6.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.5
binaryornot==0.4.4
bleach==3.1.4
blinker==1.4
cairocffi==0.9.0
CairoSVG==1.0.22
cchardet==2.1.4
celery==4.1.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
citeproc-py==0.5.1
citeproc-py-styles==0.1.3
click==7.0
configparser==3.7.4
contextlib2==0.5.5
cookiecutter==1.6.0
counter-robots==2018.6
cryptography==2.6.1
datacite==1.0.1
dcxml==0.1.1
decorator==4.4.0
defusedxml==0.5.0
dictdiffer==0.8.0
dojson==1.4.0
dulwich==0.19.11
elasticsearch==7.0.5
elasticsearch-dsl==7.0.0
entrypoints==0.3
enum34==1.1.6
Flask==1.0.4
Flask-Admin==1.5.6
Flask-Alembic==2.0.1
Flask-Assets==0.12
Flask-BabelEx==0.9.4
Flask-Breadcrumbs==0.4.0
Flask-Caching==1.6.0
Flask-CeleryExt==0.3.2
Flask-Collect==1.2.2
Flask-Cors==3.0.9
Flask-DebugToolbar==0.10.1
Flask-IIIF==0.5.0
Flask-KVSession==0.6.2
Flask-Limiter==1.0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Menu==0.7.0
Flask-OAuthlib==0.9.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.7
Flask-Security==3.0.0
flask-shell-ipython==0.4.0
Flask-SQLAlchemy==2.4.4
flask-talisman==0.6.0
Flask-WTF==0.14.2
fs==0.5.4
ftfy==4.4.3
funcsigs==1.0.2
future==0.17.1
github3.py==1.0.0a4
html5lib==1.0.1
httpretty==0.9.6
humanize==0.5.1
idna==2.7
IDUtils==1.1.13
infinity==1.4
intervals==0.8.1
invenio-access==1.1.0
invenio-accounts==1.1.1
invenio-admin==1.0.1
invenio-app==1.2.2
invenio-assets==1.0.0
invenio-banners==1.0.0a0
invenio-base==1.2.3
invenio-cache==1.0.0
invenio-celery==1.0.1
invenio-communities==1.0.0a29
invenio-config==1.0.1
invenio-csl-rest==1.0.0a1
invenio-db==1.0.4
invenio-deposit==1.0.0a11
invenio-files-rest==1.0.0a23.post3
invenio-formatter==1.0.1
invenio-github==1.0.0a28
invenio-i18n==1.0.0
invenio-iiif==1.0.0a5
invenio-indexer==1.1.1
invenio-jsonschemas==1.0.0
invenio-logging==1.1.1
invenio-mail==1.0.2
invenio-marc21==1.0.0a9
invenio-oaiserver==1.1.1.post1
invenio-oauth2server==1.0.3
invenio-oauthclient==1.1.2
invenio-openaire==1.0.0a15
invenio-opendefinition==1.0.0a9
invenio-pidrelations==1.0.0a4
invenio-pidstore==1.1.1
invenio-previewer==1.0.0a11
invenio-queues==1.0.0a2
invenio-records==1.3.2
invenio-records-files==1.0.0a11.post1
invenio-records-rest==1.6.6
invenio-records-ui==1.0.1
invenio-rest==1.1.3
invenio-search==1.2.3
invenio-search-ui==1.0.1
invenio-sipstore==1.0.0a7
invenio-stats==1.0.0a14.post2
invenio-theme==1.0.0
invenio-userprofiles==1.0.1
invenio-webhooks==1.0.0a4
ipaddr==2.2.0
ipaddress==1.0.22
ipython==5.8.0
ipython-genutils==0.2.0
isbnlib==3.10.10
itsdangerous==1.1.0
Jinja2==2.10.1
jinja2-time==0.2.0
joblib==0.14.1
jsmin==2.2.2
jsonpatch==1.23
jsonpointer==2.0
jsonref==0.2
jsonresolver==0.2.1
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kombu==4.4.0
limits==1.3
lxml==4.6.2
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.20.5
maxminddb==1.5.2
maxminddb-geolite2==2018.703
mistune==0.8.4
mock==2.0.0
msgpack-python==0.5.6
nbconvert==5.4.1
nbformat==4.4.0
ndg-httpsclient==0.4.4
node-semver==0.1.1
numpy==1.16.6
oauthlib==2.1.0
pandocfilters==1.4.2
passlib==1.7.1
pathlib2==2.3.3
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.2.2
pluggy==0.12
poyo==0.4.2
prompt-toolkit==1.0.15
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pyasn1==0.4.2
pycountry==18.12.8
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pykerberos==1.2.1
pyOpenSSL==17.5.0
python-dateutil==2.8.0
python-editor==1.0.4
python-geoip==1.2
python-slugify==3.0.1
pytz==2018.9
PyYAML==5.4.1
pyzmq==18.0.1
raven==6.10.0
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.1.0
requests-kerberos==0.12.0
scandir==1.10.0
scikit-learn==0.20.4
scipy==1.2.3
Sickle==0.6.4
simplegeneric==0.8.1
simplejson==3.16.0
simplekv==0.12.0
singledispatch==3.4.0.3
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.24
SQLAlchemy-Continuum==1.3.6
SQLAlchemy-Utils==0.33.11
statsd==3.3.0
testpath==0.4.2
text-unidecode==1.2
tornado==5.1.1
traitlets==4.3.2
ua-parser==0.8.0
Unidecode==1.0.22
uritemplate.py==1.0.1
uritools==2.2.0
urllib3==1.24.2
uWSGI==2.0.20
uwsgitop==0.11
validators==0.12.6
vine==1.3.0
Wand==0.5.2
wcwidth==0.1.7
webargs==5.5.1
webassets==0.12.1
webencodings==0.5.1
Werkzeug==0.15.4
whichcraft==0.5.2
wsgi-statsd==0.3.1
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.3
